swagger eklemesi ve model updatlerı

+++ UserManagmentAPI.DataAccess/Migrations/20210420073734_mig.cs
@@ -11,7 +11,7 @@ namespace UserManagmentAPI.DataAccess.Migrations
                 name: "Users",
                 columns: table => new
                 {
-                    ID = table.Column<int>(type: "int", nullable: false)
+                    ID = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(50)", maxLength: 50, nullable: true),
                     Surname = table.Column<string>(type: "nvarchar(50)", maxLength: 50, nullable: true),
